+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 -   linting the code automatically via git hook on commit action
 -   `<HandleTools />`: can be used as single handle content to add an context menu to handles
+-   `<NodeContent />`
+    -   `introductionTime` parameter could be used to visualize the node was added or updated
 -   `<SuggestField />`
     -   will replace `<AutoCompleteField />`
     -   match dropdown to element width when `fill=true`